requirements.txt
python-dotenv>=1.2
DrissionPage>=4.1
requests>=2.32